server/requirements.txt
aiohttp==3.7.3
appnope==0.1.2
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
blessings==1.7
bpython==0.20.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
curtsies==0.3.4
cycler==0.10.0
decorator==5.0.7
dnspython==2.1.0
dpkt==1.9.5
greenlet==1.0.0
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.2
lyricsgenius==3.0.0
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
pandas==1.2.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
progressbar==2.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
pypi-install==0.0.5
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scapy==2.4.4
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
slack==0.0.2
slackclient==2.9.3
slacker==0.14.0
soupsieve==2.2.1
spotipy==2.17.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
wordcloud==1.8.1
yarl==1.6.3
Flask==1.1.2
gunicorn==20.1.0
nltk==3.6.2
vaderSentiment==3.3.2